updated to float in db mifration and model

+++ api/src/db/migrations/2.5-create-work.js
@@ -16,7 +16,7 @@ module.exports = {
                 type: Sequelize.STRING,
             },
             rate: {
-                type: Sequelize.FLOAT(3, 2),
+                type: Sequelize.FLOAT,
                 defaultValue: 0.0,
             },
             workStatusId: {
